requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
chromadb>=0.4.22
pypdf>=3.17.0
python-dotenv>=1.0.0
openai>=1.12.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
